services/product-service/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
python-dotenv==1.0.0
dj-database-url==2.1.0
Pillow==10.2.0
pika==1.3.2
